+++ src/modules/cityData.js
@@ -1,0 +1,14 @@
+import { cityURL } from './apiURL';
+
+const fetchCityData = async () => {
+  const response = await fetch(cityURL)
+    .then((res) => res.json())
+    .then((data) => {
+      console.log(data);
+      const { country, lat, lon } = data[0];
+      return { country, lat, lon };
+    });
+  return response;
+};
+
+export default fetchCityData;
